+++ pkg/manifest/script.go
@@ -176,25 +176,28 @@ func GenerateTestCases(params *GenerationParameters) ([]model.TestCase, Scripts,
 }
 
 func addConditionalPropertiesToRequest(tc *model.TestCase, conditional []discovery.ConditionalAPIProperties, log *logrus.Entry) error {
-
 	for _, cond := range conditional {
 		for _, ep := range cond.Endpoints {
 			if tc.Input.Method == ep.Method && tc.Input.Endpoint == ep.Path {
 				// try to add property to body request
 				for _, prop := range ep.ConditionalProperties {
-					isRequestProperty, err := tc.Validator.IsRequestProperty(tc.Input.Method, tc.Input.Endpoint, prop.Path)
+					isRequestProperty, propertyType, err := tc.Validator.IsRequestProperty(tc.Input.Method, tc.Input.Endpoint, prop.Path)
 					if err != nil {
 						log.Error(err)
 						return err
 					}
 					if isRequestProperty && len(prop.Value) > 0 {
 						var err error
-						tc.Input.RequestBody, err = sjson.Set(tc.Input.RequestBody, prop.Path, prop.Value)
+						if propertyType == "[array]" {
+							stringArray := convertInputStringToArray(prop.Value)
+							tc.Input.RequestBody, err = sjson.Set(tc.Input.RequestBody, prop.Path, stringArray)
+						} else {
+							tc.Input.RequestBody, err = sjson.Set(tc.Input.RequestBody, prop.Path, prop.Value)
+						}
 						if err != nil {
 							log.Error(err)
 							return err
 						}
-						log.Tracef("Conditional body set to : %s", tc.Body)
 					}
 				}
 			}
@@ -202,6 +205,10 @@ func addConditionalPropertiesToRequest(tc *model.TestCase, conditional []discove
 	}
 
 	return nil
+}
+
+func convertInputStringToArray(value string) []string {
+	return strings.Split(value, ",")
 }
 
 var fnReplacementRegex = regexp.MustCompile(`[^\$fn:]?\$fn:([\w|_]*)\(([\w,\s-]*)\)`)